+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob/master/CHANGES)._
 
 #### 1.9.11-SNAPSHOT
+``
+- **[FEATURE]** (`ktx-scene2d`) Added `image` builders for `NinePatch`, `TextureRegion`, `Texture` and `Drawable`.
 
 #### 1.9.11-b1
 
@@ -8,7 +10,6 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 - **[UPDATE]** Updated to Kotlin Coroutines 1.3.6.
 - **[UPDATE]** Updated VisUI to 1.4.6.
 - **[FEATURE]** (`ktx-scene2d`) Added a generic `container` factory method that supports adding a custom actor.
-- **[FEATURE]** (`ktx-scene2d`) Added `image` builders for NinePatch, TextureRegion, Texture and Drawable
 
 #### 1.9.10-b7