+++ CHANGELOG.md
@@ -2,6 +2,8 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 
 #### 1.9.11-SNAPSHOT
 ``
+- **[UPDATE]** Updated to Kotlin 1.4.0.
+- **[UPDATE]** Updated to Kotlin Coroutines 1.3.9.
 - **[FEATURE]** (`ktx-scene2d`) Added `image` builders for `NinePatch`, `TextureRegion`, `Texture` and `Drawable`.
 
 #### 1.9.11-b1