+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob/master/CHANGES)._
 
 #### 1.9.11-SNAPSHOT
-``
+
 - **[UPDATE]** Updated to Kotlin 1.4.10.
 - **[UPDATE]** Updated to Kotlin Coroutines 1.4.0.
 - **[UPDATE]** Updated to Gradle 6.7.